Update `get_features` function
When get features with feature index file, previous function performs once load all features then output with masking according to he feature index file. This will be efficient for computation time, but this is not efficient for memory. One motivation to use feature index file is quick check the decoding performance for very huge number of units in that layer (such as video model). In this case, It will be better to take memory efficient rather than computation time. I also feel even if it is efficient for computation time, it is useless the memory is over.